requirements.txt
langchain
langchain-openai
openai
chromadb
pypdf
tiktoken
python-dotenv
streamlit